Fix copy CSS button on Safari
Why is it always Safari.

Anyways, for reasons I don't understand, having the fieldset have a display:
flex was causing the "copy" button its legend to not work on the generated CSS
section. So we instead selectively add the display: flex to the other two
fieldsets, and for this one, reduce the width of the textarea to 99% to make it
fit (instead of relying on the display: flex).